backend/requirements.txt
aniso8601==7.0.0
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
beautifulsoup4==4.6.0
captcha==0.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.6.0
django-crispy-forms==1.9.0
django-extensions==3.1.0
django-graphql-jwt==0.3.1
django-heroku==0.3.1
django-toggle-switch-widget==0.1.0
entrypoints==0.3
filelock==3.0.12
graphene==2.1.8
graphene-django==2.13.0
graphene-file-upload==1.2.2
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
isort==4.3.21
jedi==0.15.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.6.1
pipenv==2018.11.26
pluggy==0.13.1
promise==2.3
psycopg2==2.8.5
pycodestyle==2.5.0
pydocstyle==5.0.2
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.4.4
python-jsonrpc-server==0.3.4
python-language-server==0.31.8
pytz==2019.3
rope==0.16.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
snowballstemmer==2.0.0
sqlparse==0.3.0
Unidecode==1.1.2
virtualenv==20.0.7
virtualenv-clone==0.5.3
whitenoise==5.0.1
wrapt==1.11.2
yapf==0.29.0



